requirements.txt
fire~=0.4.0
Pillow~=9.2.0